consumer/requirements.txt
fastapi
uvicorn
kafka-python
elasticsearch==8.14.0
requests